ai_tools/predictive_analytics.py
from django.db import models
from agents.models import Agent
from .models import PredictionPerformance
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler
import numpy as np
import pandas as pd

class PredictiveAnalytics:
    def __init__(self):
        self.model = RandomForestRegressor()
        self.scaler = StandardScaler()

    def prepare_data(self, agent_id):
        """Prépare les données historiques pour l'entraînement"""
        # TODO: Implémenter la récupération et préparation des données
        pass
        
    def train_model(self, X, y):
        """Entraîne le modèle sur les données historiques"""
        self.model.fit(X, y)
        
    def predict_performance(self, agent_id):
        """Prédit les performances futures d'un agent"""
        # Version simplifiée pour le moment
        prediction = np.random.uniform(0.6, 0.9)
        confidence = np.random.uniform(0.7, 0.9)
        
        # Sauvegarder la prédiction
        PredictionPerformance.objects.create(
            agent_id=agent_id,
            kpi_predit=prediction,
            confiance=confidence,
            alerte=(prediction < 0.7)  # Alerte si KPI prédit < 70%
        )
        
        return {
            'prediction': prediction,
            'confidence': confidence,
            'alert': prediction < 0.7
        }
        
    def detect_performance_drop(self, agent_id, threshold=0.15):
        """Détecte les baisses significatives de performance"""
        # TODO: Implémenter la détection
        pass
