requirements.txt
django
mysqlclient
django-widget-tweaks
djangorestframework

# AI Dependencies
scikit-learn
tensorflow
transformers
ortools
pandas
numpy
textblob
celery
redis
django-celery-beat
django-celery-results
